Add state change callback.
